requirements.txt
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.18.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
platformdirs==4.3.7
pycparser==2.22
PyMuPDF==1.25.5
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pyzbar==0.1.9
six==1.17.0
tzdata==2025.2
virtualenv==20.30.0
